+++ catscripts/Government/Countries/Lesotho/Lesotho.sh
@@ -1,14 +1,31 @@
 #!/bin/bash
 
-egrep -i 'Lesotho' newpages.txt >> Lesotho.txt
+KEYWORDS_LESOTHO="Lesotho"
+KEYWORDS_LESOTHO_ALL="$KEYWORDS_LESOTHO"
 
-LESOTHO=`stat --print=%s Lesotho.txt`
-
-if [ $LESOTHO -ne 0 ];
+if [ "$1" == "" ]; #Normal operation
 then
-  export CATFILE="Lesotho.txt"
-  export CATNAME="Lesotho"
-  $CATEGORIZE
-fi
+  
+  if [ "$DEBUG" == "yes" ]; 
+  then
+    printf "Starting Lesotho\n"
+  fi
 
-rm Lesotho.txt
+  LESOTHO=`egrep -i "$KEYWORDS_LESOTHO" newpages.txt`
+
+  if [ "$LESOTHO" != "" ];
+  then
+    printf "$LESOTHO" > Lesotho.txt
+    export CATFILE="Lesotho.txt"
+    export CATNAME="Lesotho"
+    $CATEGORIZE
+    rm Lesotho.txt
+    unset LESOTHO
+  fi
+
+  if [ "$DEBUG" == "yes" ]; 
+  then
+    printf "Finishing Lesotho\n"
+  fi
+
+fi
