+++ catscripts/Maintenance/Noninterrogative.sh
@@ -1,14 +1,61 @@
 #!/bin/bash
 
-egrep -iv "\bwho(|s)\b" newpages.txt| egrep -iv "\bwhat" | egrep -iv "\bwhere\b" | egrep -iv "\bwhen\b" | egrep -iv "(^|\[\[|(\.|,) )Whens\b" | egrep -iv "\bwhy" | egrep -iv "\bhow\b" | egrep -iv "(^|\[\[|(\.|,) )Do " | egrep -iv "(^|\[\[|(\.|,) )has\b" | egrep -iv "(^|\[\[|\. |, )does\b" | egrep -iv "(^|\[\[|\. |, )did\b" | egrep -iv "(^|\[\[|\. |, )Will\b" | egrep -iv "(^|\[\[|(\.|,) )Is\b" | egrep -iv "(^|\[\[|(\.|,) )Are\b" | egrep -iv "(^|\[\[|(\.|,) )Aren't\b" | egrep -iv "(^|\[\[|(\.|,) )were\b" | egrep -iv '(^|\[\[|(\.|,) )was\b' | egrep -iv "(^|\[\[|(\.|,) )which\b" | egrep -iv "(^|\[\[|(\.|,) )could" | egrep -iv "(^|\[\[|(\.|,) )should\b" | egrep -iv "(^|\[\[|(\.|,) )shouldn(|')t" | egrep -iv "(^|\[\[|(\.|,) )would" | egrep -iv "(^|\[\[|\. |, )can\b" | egrep -iv "(^|\[\[|(\.|,) )Has\b" | egrep -iv "(^|\[\[|(\.|,) )Hasn't" | egrep -iv "\bAm I\b" | egrep -iv "(^|\[\[|(\.|,) )Have(|n(|')t)\b" | egrep -iv "(^|\[\[|(\.|,) )(Isn(|')t)" | egrep -iv "(^|\[\[|(\.|,) )Doesn(|')t\b" | egrep -iv "(^|\[\[|(\.|,) )Can('|)t" | egrep -iv "(^|\[\[|(\.|,) )Didn(|')t" >> noninterrogative.txt
 
-NONINT=`stat --print=%s noninterrogative.txt`
+KEYWORDS_WHO="\bwho(|s)\b"
+KEYWORDS_WHAT="\bwhat"
+KEYWORDS_WHERE="\bwhere\b"
+KEYWORDS_WHEN="\bwhen\b|(^|\[\[|(\.|,) )Whens\b"
+KEYWORDS_WHY="\bwhy"
+KEYWORDS_HOW="\bhow\b"
+KEYWORDS_DO="(^|\[\[|(\.|,) )Do "
+KEYWORDS_DOESNT="(^|\[\[|(\.|,) )Doesn(|')t\b"
+KEYWORDS_HAS="(^|\[\[|(\.|,) )has\b"
+KEYWORDS_HASNT="(^|\[\[|(\.|,) )Hasn(|')t"
+KEYWORDS_DOES="(^|\[\[|\. |, )does\b"
+KEYWORDS_DID="(^|\[\[|\. |, )did\b"
+KEYWORDS_DIDNT="(^|\[\[|(\.|,) )Didn(|')t"
+KEYWORDS_WILL="(^|\[\[|\. |, )Will\b"
+KEYWORDS_IS="(^|\[\[|(\.|,) )Is\b"
+KEYWORDS_ISNT="(^|\[\[|(\.|,) )(Isn(|')t)"
+KEYWORDS_ARE="(^|\[\[|(\.|,) )Are\b"
+KEYWORDS_ARENT="(^|\[\[|(\.|,) )Aren(|')t\b"
+KEYWORDS_WERE="(^|\[\[|(\.|,) )were\b"
+KEYWORDS_WAS="(^|\[\[|(\.|,) )was\b"
+KEYWORDS_WHICH="(^|\[\[|(\.|,) )which\b"
+KEYWORDS_COULD="(^|\[\[|(\.|,) )could"
+KEYWORDS_SHOULD="(^|\[\[|(\.|,) )should\b"
+KEYWORDS_SHOULDNT="(^|\[\[|(\.|,) )shouldn(|')t"
+KEYWORDS_WOULD="(^|\[\[|(\.|,) )would"
+KEYWORDS_CAN="(^|\[\[|\. |, )can\b"
+KEYWORDS_CANT="(^|\[\[|(\.|,) )Can('|)t"
+KEYWORDS_AMI="\bAm I\b"
+KEYWORDS_HAVE="(^|\[\[|(\.|,) )Have(|n(|')t)\b"
 
-if [ $NONINT -ne 0 ];
+KEYWORDS_INTERROGATIVE="$KEYWORDS_WHO|$KEYWORDS_WHAT|$KEYWORDS_WHERE|$KEYWORDS_WHEN|$KEYWORDS_WHY|$KEYWORDS_HOW|$KEYWORDS_DO|$KEYWORDS_HAS|$KEYWORDS_HASNT|$KEYWORDS_DOES|$KEYWORDS_DOESNT|$KEYWORDS_DID|$KEYWORDS_DIDNT|$KEYWORDS_WILL|$KEYWORDS_IS|$KEYWORDS_ISNT|$KEYWORDS_ARE|$KEYWORDS_ARENT|$KEYWORDS_WERE|$KEYWORDS_WAS|$KEYWORDS_WHICH|$KEYWORDS_COULD|$KEYWORDS_SHOULD|$KEYWORDS_SHOULDNT|$KEYWORDS_WOULD|$KEYWORDS_CAN|$KEYWORDS_CANT|$KEYWORDS_AMI|$KEYWORDS_HAVE"
+
+
+if [ "$1" == "" ]; #Normal operation
 then
-  export CATFILE="noninterrogative.txt"
-  export CATNAME="Page titles lacking an interrogative word"
-  $CATEGORIZE
-fi
+  
+  if [ "$DEBUG" == "yes" ];
+  then
+    printf "Starting Non-interrogative question search\n"
+  fi
 
-rm noninterrogative.txt
+  NONINT=`egrep -iv "$KEYWORDS_INTERROGATIVE" newpages.txt`
+
+  if [ "$NONINT" != "" ];
+  then
+    printf "$NONINT" > noninterrogative.txt
+    export CATFILE="noninterrogative.txt"
+    export CATNAME="Page titles lacking an interrogative word"
+    $CATEGORIZE
+    rm noninterrogative.txt
+  fi
+
+  if [ "$DEBUG" == "yes" ];
+  then
+    printf "Finishing Non-interrogative question search\n"
+  fi
+
+fi
